Replace fake updates with real ones in Polling mode; add utility to confirm all pending updates for a bot

+++ src/main/scala/PPS19/scalagram/logic/Bot.scala
@@ -29,7 +29,7 @@ sealed trait Bot {
   def launch(mode: Mode): Unit = mode.start(this)
 
   def getUpdates(
-      offset: Option[Int] = None,
+      offset: Option[Long] = None,
       limit: Option[Int] = None,
       timeout: Option[Int] = None,
       allowedUpdates: Option[Array[String]] = None
